Refactor authentication-pseudo MVC design pattern
